% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_aggregates.R
\name{multi_trace_plotly}
\alias{multi_trace_plotly}
\title{Plot multi-trace subplots with plotly.}
\usage{
multi_trace_plotly(
  df,
  in_var,
  de_var,
  groupby_var = NULL,
  plot_type = list(type = "scatter", mode = "lines"),
  orientation = c("v", "h"),
  return_type = c("subplot", "list")
)
}
\arguments{
\item{df}{A data.frame.}

\item{in_var}{A character. Character of the column to be used for independent
variable, typically x axis in "vertical" orientation.}

\item{de_var}{A character, or (named) charactor vector, or (named) list of
character vector. The structure will determine how the dependent variables
are arranged into multi-axis subplots, typically y axis in "vertical"
orientation.}

\item{groupby_var}{A character, or character vector. Character of the
column(s) to be used for grouping, currently implemented by linetype.}

\item{plot_type}{Named list for plotly type, mode, etc.}

\item{orientation}{"v" or "h", for "vertical" and "horizontal". v (vertical)
is the common x/y arrangement e.g. top edge, h (horizontal) is the uncommon
 x/y arrangement e.g. right edge.}

\item{return_type}{"subplot" or "list". return_type should be subplot or list
of plots.}
}
\value{
A plotly (subplot) object or a list of plotly objects depending on
  return_type.
}
\description{
\code{multi_trace_plotly} plots multi-trace subplots with plotly.
}
\details{
This is a function to plot multi-trace subplots using plotly package.
Multiple dependent variables are plot agains independent variable. Dependent
variables can be organized/combined in different subplots. Groupby variable
and plot types can be controlled. The plots can be both vertical (default)
and horizontal orientations. It can return either subplots or a list of
plotly objects.
}
\examples{
\dontrun{
multi_trace_plotly(df, in_var = "x_column",
  de_var = c("y_column1", "y_column2"),
  groupby_var = "z_column",
  plot_type = list(type = "scatter", mode = "lines"),
  orientation = "v", return_type = "subplot")
}

}
