% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_aggregates.R
\name{heatmap_plotly}
\alias{heatmap_plotly}
\title{Plot heatmap with plotly.}
\usage{
heatmap_plotly(df, x, y, z)
}
\arguments{
\item{df}{A data.frame.}

\item{x}{A character. Character of the column to be used for x axis.}

\item{y}{A character. Character of the column to be used for y axis.}

\item{z}{A character. Character of the column to be used for color axis.}
}
\value{
A plotly object of heatmap type.
}
\description{
\code{heatmap_plotly} plots heatmap with plotly.
}
\details{
This is a function to plot heatmap using plotly package. Sensible defaults
are supplied to make the plot better conform with desired appearance.
}
\examples{
\dontrun{
heatmap_plotly(df, "x_column", "y_column", "z_column")
}

}
